♻️ : #467 - feedback/evaluate 비즈니스 로직 분리
