+++ src/app/feedback/evaluate/_components/hooks/index.ts
@@ -1,1 +1,2 @@
 export { default as useEvaluateRadio } from './useEvaluateRadio';
+export { default as useEvaluatePage } from './useEvaluatePage';
